+++ editor.c
@@ -1,5 +1,4 @@
 #include <string.h>
-#include <stdint.h>
 
 #include "editor.h"
 #include "config.h"
@@ -24,7 +23,7 @@ void move_up(Buffer *b) {
 
 void move_right(Buffer *b) {
     char *lp = line_goto(b->data, b->line);
-    uint32_t len = line_len(lp);
+    int len = line_len(lp);
     if (b->col < len) {
         b->col++;
         b->col_max = b->col;
@@ -99,9 +98,9 @@ void move_backward(Buffer *b) {
 
 void move_line_begin(Buffer *b) {
     char *cur_line = line_goto(b->data, b->line);
-    uint32_t len = line_len(cur_line);
+    int len = line_len(cur_line);
 
-    for (size_t i = 0; i < len; ++i) {
+    for (int i = 0; i < len; ++i) {
         if (cur_line[i] != ' ') {
             b->col = b->col_max = i;
             break;
@@ -111,36 +110,36 @@ void move_line_begin(Buffer *b) {
 
 void move_line_end(Buffer *b) {
     char *cur_line = line_goto(b->data, b->line);
-    uint32_t len = line_len(cur_line);
+    int len = line_len(cur_line);
     b->col = b->col_max = len;
 }
 
 // TODO rename
-void move_screen_down(Buffer *b, uint16_t screen_height) {
-    uint16_t n = screen_height / 2;
-    uint32_t pos = b->line + n;
+void move_screen_down(Buffer *b, int screen_height) {
+    int n = screen_height / 2;
+    int pos = b->line + n;
     b->line = (pos > line_count(b->data) - 1) ? line_count(b->data) - 1: pos;
     adjust_col(b);
 }
 
 // TODO rename
-void move_screen_up(Buffer *b, uint16_t screen_height) {
-    uint16_t n = screen_height / 2;
-    int32_t pos = b->line - n;
+void move_screen_up(Buffer *b, int screen_height) {
+    int n = screen_height / 2;
+    int pos = b->line - n;
     b->line = (pos >= 0) ? pos : 0;
     adjust_col(b);
 }
 
 // TODO rename
-void move_screen_center(Buffer *b, uint16_t screen_height) {
-    int32_t off = b->line - (screen_height / 2);
+void move_screen_center(Buffer *b, int screen_height) {
+    int off = b->line - (screen_height / 2);
     b->line_off = (off >= 0) ? off : 0;
     adjust_col(b);
 }
 
 void insert_char(Buffer *b, char c) {
-    static uint8_t size = 0;
-    static uint8_t acc = 0;
+    static int size = 0;
+    static int acc = 0;
     static char buf[4] = {0};
 
     if (char_size(c) > 0) {
@@ -152,7 +151,7 @@ void insert_char(Buffer *b, char c) {
     buf[acc++] = c;
 
     if (acc != 0 && acc == size) {
-        size_t pos = get_current_pos(b);
+        int pos = get_current_pos(b);
         insert_substr(b, pos, buf, size);
 
         if (buf[0] == '\n') {
@@ -168,8 +167,8 @@ void insert_char(Buffer *b, char c) {
 }
 
 void delete_char(Buffer *b) {
-    size_t pos = 0;
-    uint32_t prev_line_len = 0;
+    int pos = 0;
+    int prev_line_len = 0;
 
     if ((pos = get_current_pos(b)) == 0) return;
     pos--;
@@ -200,9 +199,9 @@ void clear_killed(void) {
 
 void kill_line(Buffer *b) {
     char *cur_line = line_goto(b->data, b->line);
-    uint32_t size = line_size(cur_line);
-    size_t pos = cur_line - b->data;
-    uint32_t killed = (cur_line[size] != '\0') ? size + 1 : size;
+    int size = line_size(cur_line);
+    int pos = cur_line - b->data;
+    int killed = (cur_line[size] != '\0') ? size + 1 : size;
 
     strncat(kill_buffer, cur_line, killed);
     erase_substr(b, pos, killed);
@@ -241,7 +240,7 @@ void end_region(Buffer *b) {
 void kill_region(Buffer *b) {
     if (b->reg_begin == b->reg_end) return;
 
-    uint32_t reg_len = b->reg_end - b->reg_begin;
+    int reg_len = b->reg_end - b->reg_begin;
 
     strncat(kill_buffer, b->reg_begin, reg_len);
     erase_substr(b, b->reg_begin - b->data, reg_len);
@@ -256,15 +255,15 @@ void kill_region(Buffer *b) {
 void copy_region(Buffer *b) {
     if (b->reg_begin == b->reg_end) return;
 
-    uint32_t reg_len = b->reg_end - b->reg_begin;
+    int reg_len = b->reg_end - b->reg_begin;
 
     strncat(kill_buffer, b->reg_begin, reg_len);
 }
 
 void paste(Buffer *b) {
-    uint32_t len = strlen(kill_buffer);
-    size_t pos = get_current_pos(b);
-    uint32_t count = line_count(kill_buffer);
+    int len = strlen(kill_buffer);
+    int pos = get_current_pos(b);
+    int count = line_count(kill_buffer);
 
     if (len == 0) return;
